+++ core-ktx/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## [1.6.0-1] (2021-10-03)
+
 ### Added
 
 - **Canvas**: new extensions `withClipOut`
@@ -13,6 +15,10 @@
 
 - Added `inline` modifier to Keyboard extensions
 - Default value in `SharedPreferences` delegate should not be evaluated before it needed (similar to #31)
+
+### Dependencies
+
+- kotlin-stdlib 1.5.20 -> 1.5.31
 
 ## [1.6.0-0] (2021-07-05)
 
@@ -38,5 +44,6 @@
 First release
 
 
+[1.6.0-1]: https://github.com/RedMadRobot/redmadrobot-android-ktx/compare/resources-ktx-v1.3.1-0...core-ktx-v1.6.0-1
 [1.6.0-0]: https://github.com/RedMadRobot/redmadrobot-android-ktx/compare/viewbinding-ktx-v4.2.1-0...core-ktx-v1.6.0-0
 [1.5.0-0]: https://github.com/RedMadRobot/redmadrobot-android-ktx/compare/e6b11af4...core-ktx-v1.5.0-0